move login Errors enum to login presenter
